+++ src/domains/card-list/components/card-register-button.tsx
@@ -1,24 +1,6 @@
-import { Flex, Header } from '@/components'
+import { Flex } from '@/components'
 import { IconPlus } from '@tabler/icons-react'
 import { Link } from 'react-router-dom'
-
-export const CardListPage = () => {
-  return (
-    <Flex direction="column" height="100vh" alignItems="center" justifyContent="center">
-      <Header title="보유카드" />
-      <Flex
-        as="main"
-        width="60vw"
-        height="100%"
-        direction="column"
-        gap="24px"
-        justifyContent="center"
-      >
-        <CardRegisterButton />
-      </Flex>
-    </Flex>
-  )
-}
 
 export const CardRegisterButton = () => {
   return (
