+++ src/domains/card-register/funnels/card-register-complete-funnel.tsx
@@ -1,0 +1,5 @@
+import { Text } from '@/components'
+
+export const CardRegisterCompleteFunnel = () => {
+  return <Text variant="title1">카드등록이 완료되었습니다.</Text>
+}
